+++ db-migration-dm/src/main/java/com/github/mengweijin/util/ReflectUtils.java
@@ -1,8 +1,14 @@
 package com.github.mengweijin.util;
 
+import com.zaxxer.hikari.pool.HikariProxyConnection;
+import dm.jdbc.driver.DmdbConnection;
+
+import java.lang.reflect.AccessibleObject;
 import java.lang.reflect.Field;
 import java.lang.reflect.Modifier;
+import java.sql.Connection;
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
 
 /**
@@ -28,5 +34,63 @@ public class ReflectUtils {
             }
         }
         return list;
+    }
+
+    public static List<Field> getAllFields(Class<?> clazz) {
+        Field[] declaredFields = clazz.getDeclaredFields();
+        return new ArrayList<>(Arrays.asList(declaredFields));
+    }
+
+    public static Object getFieldValue(Object obj, Field field) {
+        if (null == field) {
+            return null;
+        }
+        if (obj instanceof Class) {
+            // 静态字段获取时对象为null
+            obj = null;
+        }
+
+        setAccessible(field);
+        Object result;
+        try {
+            result = field.get(obj);
+        } catch (IllegalAccessException e) {
+            throw new RuntimeException(e);
+        }
+        return result;
+    }
+
+    public static <T extends AccessibleObject> void setAccessible(T accessibleObject) {
+        if (null != accessibleObject && !accessibleObject.isAccessible()) {
+            accessibleObject.setAccessible(true);
+        }
+    }
+
+    public static boolean isClassExist(String className) {
+        try {
+            Class.forName(className);
+            return true;
+        } catch (ClassNotFoundException e) {
+            return false;
+        }
+    }
+
+    public static Object getHikariProxyConnectionDelegateFieldValue(Connection connection) {
+        try {
+            Field delegate = connection.getClass().getSuperclass().getDeclaredField("delegate");
+            return getFieldValue(connection, delegate);
+        } catch (NoSuchFieldException e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+    public static boolean isDmConnection(Connection connection) {
+        if(connection instanceof DmdbConnection) {
+            return true;
+        }
+        if(ReflectUtils.isClassExist("com.zaxxer.hikari.pool.HikariProxyConnection")) {
+            return connection instanceof HikariProxyConnection && getHikariProxyConnectionDelegateFieldValue(connection) instanceof DmdbConnection;
+        }
+        return false;
     }
 }