+++ demo/liquibase-dm-spring-boot2/src/main/java/com/github/mengweijin/liquibase/LiquibaseDMSpringBoot2Application.java
@@ -1,0 +1,15 @@
+package com.github.mengweijin.liquibase;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+/**
+ * @author mengweijin
+ */
+@SpringBootApplication
+public class LiquibaseDMSpringBoot2Application {
+
+	public static void main(String[] args) {
+		SpringApplication.run(LiquibaseDMSpringBoot2Application.class, args);
+	}
+}