+++ demo/liquibase-dm/src/main/java/com/github/mengweijin/liquibase/LiquibaseDMApplication.java
@@ -1,4 +1,4 @@
-package com.github.mengweijin;
+package com.github.mengweijin.liquibase;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -7,9 +7,9 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
  * @author mengweijin
  */
 @SpringBootApplication
-public class FlywayDMApplication {
+public class LiquibaseDMApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(FlywayDMApplication.class, args);
+		SpringApplication.run(LiquibaseDMApplication.class, args);
 	}
 }